Remove old runSmoking function, and move smoking functionality into its separate cpp file
